requirements.txt
pyspark>=3.4,<4.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pyyaml>=6.0
nltk>=3.8
